+++ mayan/apps/document_states/migrations/0025_alter_workflowtransitiontriggerevent_unique_together.py
@@ -4,7 +4,10 @@ from django.db import migrations
 class Migration(migrations.Migration):
     dependencies = [
         ('events', '0008_auto_20180315_0029'),
-        ('document_states', '0024_auto_20220124_1257')
+        (
+            'document_states',
+            '0024_remove_duplicate_workflow_template_transition_triggers'
+        )
     ]
 
     operations = [