+++ mayan/apps/documents/apps.py
@@ -218,25 +218,13 @@ class DocumentsApp(MayanAppConfig):
         super().ready()
 
         Document = self.get_model(model_name='Document')
-        DocumentSearchResult = self.get_model(
-            model_name='DocumentSearchResult'
-        )
         DocumentFile = self.get_model(model_name='DocumentFile')
         DocumentFilePage = self.get_model(model_name='DocumentFilePage')
-        DocumentFileSearchResult = self.get_model(
-            model_name='DocumentFileSearchResult'
-        )
-        DocumentFilePageSearchResult = self.get_model(
-            model_name='DocumentFilePageSearchResult'
-        )
         DocumentType = self.get_model(model_name='DocumentType')
         DocumentTypeFilename = self.get_model(
             model_name='DocumentTypeFilename'
         )
         DocumentVersion = self.get_model(model_name='DocumentVersion')
-        DocumentVersionSearchResult = self.get_model(
-            model_name='DocumentVersionSearchResult'
-        )
         DocumentVersionPage = self.get_model(model_name='DocumentVersionPage')
         DocumentVersionPageSearchResult = self.get_model(
             model_name='DocumentVersionPageSearchResult'
@@ -995,22 +983,7 @@ class DocumentsApp(MayanAppConfig):
                 link_document_multiple_restore, link_document_multiple_delete
             ), sources=(TrashedDocument,)
         )
-
-        # RecentlyAccessedDocument
-
-        menu_multi_item.add_proxy_inclusions(source=RecentlyAccessedDocument)
-
-        # RecentlyCreatedDocument
-
-        menu_multi_item.add_proxy_inclusions(source=RecentlyCreatedDocument)
-
-        # Search proxies
-
-        menu_multi_item.add_proxy_inclusions(source=DocumentSearchResult)
-        menu_multi_item.add_proxy_inclusions(source=DocumentFileSearchResult)
-        menu_multi_item.add_proxy_inclusions(source=DocumentFilePageSearchResult)
-        menu_multi_item.add_proxy_inclusions(source=DocumentVersionSearchResult)
-        menu_multi_item.add_proxy_inclusions(source=DocumentVersionPageSearchResult)
+        menu_multi_item.add_proxy_exclusion(source=TrashedDocument)
 
         post_migrate.connect(
             dispatch_uid='documents_handler_create_document_file_page_image_cache',