+++ mayan/apps/documents/serializers/document_file_serializers.py
@@ -2,7 +2,6 @@ from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.rest_api import serializers
 from mayan.apps.rest_api.relations import MultiKwargHyperlinkedIdentityField
-from mayan.apps.rest_api.serializer_mixins import CreateOnlyFieldSerializerMixin
 
 from ..literals import DOCUMENT_FILE_ACTION_PAGE_CHOICES
 from ..models.document_file_models import DocumentFile
@@ -69,11 +68,9 @@ class DocumentFilePageSerializer(serializers.HyperlinkedModelSerializer):
         )
 
 
-class DocumentFileSerializer(
-    CreateOnlyFieldSerializerMixin, serializers.HyperlinkedModelSerializer
-):
+class DocumentFileSerializer(serializers.HyperlinkedModelSerializer):
     action = serializers.ChoiceField(
-        choices=DOCUMENT_FILE_ACTION_PAGE_CHOICES
+        choices=DOCUMENT_FILE_ACTION_PAGE_CHOICES, write_only=True
     )
     document_url = serializers.HyperlinkedIdentityField(
         lookup_field='document_id',
@@ -95,7 +92,7 @@ class DocumentFileSerializer(
     )
     file_new = serializers.FileField(
         help_text=_('Binary content for the new file.'),
-        use_url=False
+        use_url=False, write_only=True
     )
     page_list_url = MultiKwargHyperlinkedIdentityField(
         view_kwargs=(
@@ -133,7 +130,7 @@ class DocumentFileSerializer(
         }
         fields = (
             'action', 'checksum', 'comment', 'document_id', 'document_url',
-            'download_url', 'encoding', 'file', 'filename', 'file_new', 'id',
+            'download_url', 'encoding', 'file', 'file_new', 'filename', 'id',
             'mimetype', 'page_list_url', 'pages_first', 'size', 'timestamp',
             'url'
         )