+++ mayan/apps/documents/serializers/document_serializers.py
@@ -1,7 +1,6 @@
 from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.rest_api import serializers
-from mayan.apps.rest_api.serializer_mixins import CreateOnlyFieldSerializerMixin
 from mayan.apps.storage.models import SharedUploadedFile
 
 from ..models.document_models import Document
@@ -14,9 +13,7 @@ from .document_type_serializers import DocumentTypeSerializer
 from .document_version_serializers import DocumentVersionSerializer
 
 
-class DocumentSerializer(
-    CreateOnlyFieldSerializerMixin, serializers.HyperlinkedModelSerializer
-):
+class DocumentSerializer(serializers.HyperlinkedModelSerializer):
     document_type = DocumentTypeSerializer(read_only=True)
     document_type_id = serializers.IntegerField(
         help_text=_('Document type ID for the new document.'), write_only=True