+++ mayan/apps/messaging/serializers.py
@@ -3,7 +3,6 @@ from django.utils.translation import ugettext_lazy as _
 from rest_framework.reverse import reverse
 
 from mayan.apps.rest_api import serializers
-from mayan.apps.rest_api.serializer_mixins import CreateOnlyFieldSerializerMixin
 from mayan.apps.rest_api.relations import FilteredPrimaryKeyRelatedField
 from mayan.apps.user_management.querysets import get_user_queryset
 from mayan.apps.views.utils import resolve
@@ -11,9 +10,7 @@ from mayan.apps.views.utils import resolve
 from .models import Message
 
 
-class MessageSerializer(
-    CreateOnlyFieldSerializerMixin, serializers.HyperlinkedModelSerializer
-):
+class MessageSerializer(serializers.HyperlinkedModelSerializer):
     sender_app_label = serializers.SerializerMethodField()
     sender_model_name = serializers.SerializerMethodField()
     sender_url = serializers.SerializerMethodField()