+++ mayan/apps/user_management/tests/test_group_views.py
@@ -5,9 +5,7 @@ from mayan.apps.metadata.permissions import permission_document_metadata_edit
 from mayan.apps.metadata.tests.mixins import MetadataTypeTestMixin
 from mayan.apps.testing.tests.base import GenericViewTestCase
 
-from ..events import (
-    event_group_created, event_group_edited, event_user_edited
-)
+from ..events import event_group_created, event_group_edited
 from ..permissions import (
     permission_group_create, permission_group_delete, permission_group_edit,
     permission_group_view, permission_user_edit