+++ mayan/settings/testing/gitlab-ci.py
@@ -1,0 +1,4 @@
+from .base import *  # NOQA
+
+SIGNATURES_GPG_PATH = '/usr/bin/gpg1'
+